fix: fix some codereview
