+++ data-agent-backend/src/main/java/io/github/malonetalk/dto/datasource/ColumnInfo.java
@@ -15,8 +15,9 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  * limitations under the License.
  */
-package io.github.malonetalk.agent.datasource;
+package io.github.malonetalk.dto.datasource;
 
+/** 物理数据源列信息，由 SchemaReader 从 JDBC 元数据读取 */
 public record ColumnInfo(
         String columnName,
         String typeName,
