+++ data-agent-backend/src/main/java/io/github/malonetalk/service/semantic/table/TableSemanticService.java
@@ -17,8 +17,8 @@
  */
 package io.github.malonetalk.service.semantic.table;
 
-import io.github.malonetalk.agent.tools.response.TablePromptResponse;
 import io.github.malonetalk.dto.pagination.PageResponse;
+import io.github.malonetalk.dto.prompt.TablePromptResponse;
 import io.github.malonetalk.dto.semantic.TableSemanticPageQuery;
 import io.github.malonetalk.dto.semantic.TableSemanticResponse;
 import io.github.malonetalk.dto.semantic.TableSemanticUpdateRequest;
